Chatbot/requirement.txt
streamlit
langchain
langchain-core
langchain-groq
python-dotenv
httpx
pydantic
